+++ modules/config/antd/index.d.ts
@@ -1,3 +1,3 @@
-import {BasicConfig} from "../..";
-declare const AntdConfig: BasicConfig;
-export default AntdConfig;
+import { BasicConfig } from "../.."
+declare const AntdConfig: BasicConfig
+export default AntdConfig