+++ modules/config/material/index.d.ts
@@ -1,3 +1,3 @@
-import {BasicConfig} from "../..";
-declare const MaterialConfig: BasicConfig;
-export default MaterialConfig;
+import { BasicConfig } from "../.."
+declare const MaterialConfig: BasicConfig
+export default MaterialConfig