+++ scripts/for-npm.sh
@@ -1,5 +1,5 @@
 npm install -g json
 json -I -f package.json \
   -e 'this.name="react-awesome-query-builder"' \
-  -e 'this.repository.url="https://github.com/ukrbublik/react-awesome-query-builder.git"' \
+  -e 'this.repository.url="https://github.com/phiduy/react-awesome-query-builder.git"' \
   -e 'delete this.publishConfig'