+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+- 4.5.1
+  - Migrate Material widgets to v5
+  - Fixed issue with slider styles
 - 4.5.0
   - Added basic support of export to ElasticSearch (PR #469)
   - Export all helper funcs from configUtils (PR #493)